Build httpServer sample app with other sample apps
This app requires lighttpd sources and was so far not
build automatically.
The 3rdParty/lighttpd changes are provided as part of
a manifest change, at version 1.4.49, and this uses
externalBuild to package the app instead of the previous
Makefile-based integration.

Also fixing some parsing error in the instapp.py script.

Resolves: LE-9463 LE-7123 LE-6033
Change-Id: I2bee8f9be254239d23bc91ca62ef70ac821bed88