Fix SMACK64EXEC issue for Legato R/O and DISABLE_SMACK=1
For Legato R/O compiled with DISABLE_SMACK=1, the flag
/legato/SMACK_DISABLED is no more created inside the squashfs
image.
If SMACK64EXEC attributes are set externally to several apps,
the apps won't start due to .so missing path. The ld.so won't
use the $ORIG/../lib path given by -rpath if SMACK64EXEC
attribute is set. In that case, all files libComponent*.so
need to be deployed into the default directory
/legato/systems/current/lib.

Resolves: LE-9839
Change-Id: I363441bb40a1c9033c21334952cebcb41460b850