[SMACK] Update whole tree with right SMACK label
If a Legato compiled with DISABLE_SMACK=1 is installed by update
system.target.update over a previous Legato compiled without, some
SMACK incoherencies appear. The new Legato checks for the current
SMACK label of "/legato" and if it differs from the expected label,
the whole tree hierarchy is updated with the right SMACK label.

Resolves: LE-10021
Change-Id: Ieb181184ec6bed4ad089edb127a36a74ba4a126c