[Docs] Point to https://legato.io & https://docs.legato.io
The legato.io website is now served as HTTPS, so pointing it
instead of HTTP.

Documentations is also being hosted at https://docs.legato.io/
instead of https://legato.io/legato-docs/ as it is more concise.

In both cases, existing links pointing to the previous location
are being redirected to the new one.

Resolves: LE-10094
Change-Id: I13f23b7022825ad9e984b58bc739acdf3e198d00