Mangle parameters which are C/C++ keywords.
Legato API files do not share the same set of keywords as C/C++, so
an API file can contain a parameter name which is a keyword in C/C++.
If it does, add an underscore (_) to the beginning when generating
the parameter name.  This will always work as underscores are not valid
starting characters for parameter names in Legato API files.

No mangling is required for type names as type names are already mangled
in such a way so they do not conflict with keywords.

Resolves: LE-10148
Change-Id: I3b9cbc052b6532850160d62ad494f251cbe6bade