Delay connection request between data connection service and the modem
This is workaround solution to manage a bug on the QTI side where the dhcp
client is not cleaned up properly when a session is disconnected/torn down.
Introducing this delay should give QTI enough time to complete its clean up
process properly before we initialize another connection.

Resolves: LE-10217
Related: LE-10182
Change-Id: I4026f9385a2a0a2b8be6267f3a117f292647cbaf