Remove EINTR handling from close()
Man says:
       The EINTR error is a somewhat special case.  Regarding the EINTR
error, POSIX.1-2013 says:

       If close() is interrupted by a signal that is to be
caught, it shall return -1 with errno set to EINTR and the state of
fildes is unspecified.

       This permits the behavior that occurs on Linux and many other
implementations, where, as with other errors that may be reported by
close(), the file descriptor is guaranteed to be closed.   However,
       it also permits another possibility: that the implementation
returns an EINTR error and keeps the file descriptor open.  (According
to its documentation, HP-UX's close() does this.)  The caller must
       then once more use close() to close the file descriptor, to avoid
file descriptor leaks.  This divergence in implementation behaviors
provides a difficult hurdle for portable applications, since  on
       many  implementations, close() must not be called again after an
EINTR error, and on at least one, close() must be called again.  There
are plans to address this conundrum for the next major release
       of the POSIX.1 standard.

So in the linux implementation of close(), since 'the file descriptor is
guaranteed to be closed', we should not be calling close() again.

More info on this topic:
- http://alobbs.com/post/54503240599/close-and-eintr
- https://stackoverflow.com/questions/31527706/proper-error-handling-for-fclose-impossible-according-to-manpage

Resolves: LE-5828 LE-8120
Change-Id: I681755690358493ef72105d2952c2206583f8018