[KernelModules] Support for loading kernel modules in alphabetical order
This patch provides support for backward compatibility of loading kernel
modules in alphabetical order of module names if explicit dependencies
are not listed.

A doubly linked list is used to capture the kernel modules in
alphabetical order of the module names. It leverages the fact that the
modules appearing in the config tree are already sorted alphabetically.

Resolves: LE-10184
Change-Id: If8118edf15d8f277408b8002d41ce7cea8174f92