[mkTools] Update precompiled header management for sccache
The Compile rule uses an -include location that does not exist.
Since the precompiled header is called mkTools.h.gch, this -include
statement is meant to point the compiler to the path where the
precompiled one is located, even though the $builddir/mkTools.h path
does not actually exist.

When using sccache, it eventually make sure that the -include path
exist and it fails.

It also seems that sccache does not work too well with precompiled
headers, so the usage of the .gch is not enforced.

Weirdly enough, when doing that, gcc and clang raised somes issues.
This was because the Precompile rule was not using -Wall -Werror.

Discarding deprecated warning otherwise it fails with issue:
clang-5.0: error: treating 'c-header' input as 'c++-header' when in C++ mode, this behavior is deprecated [-Werror,-Wdeprecated]

Resolves: LE-10248
Change-Id: I075fc78afb6ac43a8583e5b977a876dfdc72b3d8