Do not resume download when device is in airplane mode
In an attempt to resolve LE-9825, the changes to prevent a download from resuming
when the device is in airplane mode was reverted LE-7457. The change was reverted
to account for the use case where airplane mode was enabled but only wifi is being used.
Issue is addressed by conforming to how RTOS manages the data connection manager by
removing the dependency between the cellnet and dcs.

Resolves: LE-10309
Related: LE-10209
Change-Id: I8532df89fc483bfc9a88623b29b72ce2301b4ff3
(cherry picked from commit a9777a9e20965df15ad61a0e05015100428a841c)