[mkTools] Add parameter to set job count
Adding a -j parameter to all mktools as to set the
job counts.
By default ninja (= mktools backend) already parallelize
based on the number of CPUs, but with ccache we can raise
that number since the CPU is not really working (except for
the first build) but instead we are simply waiting for ccache
to retrieve eventual cached objects.

Resolves: LE-10248
Change-Id: I7b38d4441c727e6983a3686e5acc6f99cf2e9e46