Make further changes to route everything through the compiler cache
The initial implementation was missing some changes resulting in
parts of the builds that were not going through the cache.

CC and CXX are now being used to contain both ccache and the compiler
path, which makes it easier to have external builds transparently use
the cache. However they do need to handle the space in the ${CC}
variable properly.

Resolves: LE-10248
Change-Id: I146cfd44b35a6e74b1b351b10a2f8ba851d763ef