[Docs] Document behaviour change of le_data
A warning message has been added to the data
connection service documentation that describes
the change in behaviour. The change in behaviour
includes not turning on the radio and verifying
that the SIM is valid and unlocked.

This change has also been added to the 18.05.0 release notes

Resolves: LE-10209
Change-Id: Ic10ff37e2ee4794b5e6831cf7ff9fb4dca5512b3
(cherry picked from commit a8fd25ed90f8247b63f3d67732c91a6cd9389177)