Handle sysroot in findtoolchain script
When a Yocto toolchain is not in its default directory, the build
system is currently not able to figure out the sysroot by itself.

Asking the sysroot from the compiler (with --print-sysroot) does not
necessarily work, as Yocto >= 1.8 produces toolchains that returns
'/not/exist' instead of an actual path.

Legato could/should probably use the env. file provided with Yocto
toolchains when it is available, as some other info should be taken
from there.
This would however be a bigger change also impacting mk tools.

Resolves: LE-8640 LE-10274
Change-Id: I4618928992bbdb61c5660f04f62e1e400e663936