Fix mksys build out of 'make <target>' context
The TOOLCHAIN_DIR env variable was only declared when building through
'make wp76xx' in LEGATO_ROOT.
It was being used in the tinydtls build, resulting in a different
behavior when doing 'mksys -t wp76xx default.sdef'.

Just like for LEGATO_SYSROOT, also exporting TOOLCHAIN_DIR and TOOLCHAIN_PREFIX
in mk tools in case external builds wants to use it.

It also fixes an issue when using findtoolchain while LEGATO_ROOT is
not declared in a Yocto build context.

Resolves: LE-10340 LE-10346
Change-Id: Id4be997f1dc82db489129039e4ad19dcc00a8924