Files created at runtime in sandbox are not persisted after reboot
The files that are created in sandbox are being deleted whenever legato starts
up either through a reboot or legato restart. This issue was introduced when
trying to solve an issue where we did not want to modify the permissions of devices
whenever we added them into the sandbox.

Resolves: LE-10435
Related: LE-10385 LE-6544
Change-Id: I6488c0a6c926dc76641f24102c12e1632940933b