Fix build issues when CCACHE_PATH is set
If the environment-setup file from Yocto toolchains is sourced,
it defines a CCACHE_PATH that is restricted to the compilers
provided by the toolchain, which does not include a compiler for
the host machine.

This results in host tools being unable to build.

Unsetting that variable as to remove CCACHE_PATH from the
equation.

Resolves: LE-10323
Change-Id: I1c9298d0b18c0681b78d1b26397618059269a4ca