Import uLegato pytest framework and initial tests
This brings in the test configuration files and fixes up some issues
resulting from the new le_test implementation.

Those tests that build and run on the virtual target without
modification are also included.  Any tests that rely on Legato changes
that haven't been imported yet are excluded.

The only test app that has not been previously reviewed on the uLegato
branch is the CRC test, which was created to help illustrate the test
infrastructure in documentation.

Resolves: LE-10138
Change-Id: I473ad02dcdba60f16665580792ef63421de11249