[UnitTest] Add retry option
When defining UT_RETRY to a number > to 1 and UT_RETRY_ALL to false,
ctest will try to rerun all test that failed in the previous run.
This will save a result in a xml format for all run.
The result can be found in build/localhost/tests/unitTestResult.xml

Resolves: LE-10543
Change-Id: I74178cff396b52b9b94a115eef9d20c1020a10dd