Update daemon does not install a firmware update package
If the FW update is performed using "update" command
using the .update packege, then the firmware is
downloaded but the installation not happening.
The fix is to introduce the missing call to
le_fwupdate_Install(), hence making the logic similar
to what is utilized by fwupdateDaemon (or by AVC update).

Resolves: LE-10478
Change-Id: I657b6a01c1a1b6bbf0c393878d9681295822fe40