[kernelModules] Fix infinite loop in module load
If two or more kernel modules depend on the same module, the supervisor
will get stuck in a loop. The problem occurs because the same list node
is added to the list more than once causing a loop in the linked list.

The issue is fixed by not allowing duplicate items in the linked list
while still managing the usage count of each module correctly.

Resolves: LE-10550
Change-Id: I53cd761613a6952919a3420782a7d4c72343d774