Handle kernelroot in findtoolchain
The build of kernel modules through Yocto requires a kernel root
that contains some sources + some build scripts.

The default value for LEGATO_KERNELROOT is <sysroot>/usr/src/kernel,
which works in the Yocto SDK use case (ie, toolchain installed in /opt/swi for instance),
but does not work when building Legato through Yocto OR when building Legato next to
a Yocto build.

This tries to address this issue by updating the findtoolchain script
to handle to take care of finding a 'kernelroot'.

Resolves: LE-10580 LE-10362
Change-Id: I3c9719e464c38ccd484ba2c0469e3678926c19f9