Introduce KConfig for Legato build configuration
The KConfig settings subsume a number of disparate settings that were
previously controlled by various environment variables and preprocessor
defines.  The legacy environment variables are still used to populate
the initial values if no configuration is present, however the KConfig
output has become the single point of truth for these values going
forward.

The default behaviour of "make" and "make <target>" is preserved.  In
addition, "make menuconfig" and "make menuconfig_<target>" can be used
to explicitly change the configuration values.

The largest changes are the addition of the new KConfig files defining
the options, and extensive modifications to the Makefiles to generate
and use the configured values.  Most of the rest of the changes are
modifying preprocessor directives and values to use the new
configuration names.

Also included is updated documentation to reflect the new KConfig
options and procedures.

Resolves: LE-10145 LE-10147 LE-10146
Change-Id: I008a082e86ada648311da82c9aadf418efef0832