Coverity fix CID 157303 - tainted string
String received from getenv() is considered tainted,
as it comes from the untrusted source.
Apparently, previous fix (ensure the string does not
exceed certain size limit and is 0-terminated) is
not considered sufficient by Coverity.
Moving the "sanitizing" code to the separate function,
and more explicitly terminating string with zero.

Resolves: LE-7737
Change-Id: I01c1beb25585ef6cd454ae1ed71ebf2da3918416