[kernelModules] Add fault handling for kernel module load failure
This patch provides support for handling kernel module load failures.

Add '[optional]' tag for kernel modules to specify if the kernel module
is optional. e.g.

requires:
{
    kernelModules:
    {
        $CURDIR/drivers/example/example.mdef [optional]
    }
}

If the module is optional, the load failure is ignored and no recovery
action is taken. If the module is not optional, the recovery action is
based on if the system or an app has dependency on the failed module.

- If system has dependency on the failed module:
  Recovery action is to reboot the system.
- If an app has dependency on the failed module:
  Recovery action is to trigger the fault action specified for each
  process. The relevant faultAction for kernel modules are:
  restartApp, stopApp, reboot.

Resolves: LE-10251
Change-Id: Iaa7919304a9a247a4bb3e473b7509b5c65a70232