Coverity fix CID 157303 - tainted string
String received from getenv() is considered tainted,
as it comes from the untrusted source.
Previous fixes (ensuring the string does not
exceed certain size limit and is 0-terminated) are
not considered sufficient by Coverity, because it
also expects the string content to be inspected to ensure
there are no unwanted characters.
The fix is to use __coverity_tainted_data_sanitize__() call
to advise Coverity that the string is properly validated.

Resolves: LE-7737
Change-Id: I13cd45daa8fde4866a3412fef25235906578acdb