Add access permission capability in "requires" section of Xdefs for file and dir
Enable users to specify the access permission of the files and directories needed by the app.
When specifying the access permission for files or directories, the supervisor will ensure
that the DAC and SMACK permissions are properly set. SMACK rules are generated and set based
on the specified access permission in the Xdef files.

User will only be able to specify access permission for anything under /home/root and /mnt/flash.

In addition to all the changes related to access permission. The behaviour for requires dir
section has also changed. Instead of mounting each individual file from the dir into the app,
the directory will be mounted instead.

Resolves: LE-10457
Change-Id: I6ac486d88628cec077a9c9d384a6200b3277dba9