Add support for STRUCT type to API files.
This adds support for defining STRUCT types in API files.  The STRUCT
type is similar to C structures, and can be used in API files anywhere
a type can be used (including inside other structures).

This addition introduces a new inter-dependency in the ifgenerated APIs
in their order so that AllTypes() in ifgen/ifgen has to use a python
list than a set to serve that purpose.

Resolves: LE-10922
Change-Id: I03810d1141703022fc03a6fcda220f4d9fdb8d88