[kernelModules] Install required auto load modules at system startup
This patch provide changes to install the auto load modules always at
system startup. The previous behavior was to install the dependency auto
load module only with its parent module, i.e. if the parent module is
manual load then the required auto load module was loaded manually
with its parent module. However, the expected behaviour is to install
the auto load kernel modules always at system startup irrespective of
the parent module is auto or manual load.

This covers the scenario if multiple modules/devices not explicitly
listed as modules in the definition files depends on the same auto load
kernel module.

Resolves: LE-11113
Change-Id: I4dbc1c1564071d158cd905514be9338f64ccbdc3