Make findtoolchain compatible with Yocto 2.5
With latest version of Yocto, the build tree structure has changed
and there is no 'global sysroot' anymore.

Instead the sysroots are generated for each package builds.

This adapts the findtoolchain script to be compatible with both,
by using the sysroot generated for the legato-af package.

I'm also updating the virt script as to be able to generate a Docker
image with virt from an existing Yocto build.

Resolves: LE-11265
Change-Id: I2403192b50839817a798894f852bb495d4c38aa6