[MDC] Fix a crash when getting disconnect reason
Issue:
Calling the following APIs when there is no error during
disconnection (user-initiated disconnection for instance)
leads to a crash:
- le_mdc_GetDisconnectionReason()
- le_mdc_GetPlatformSpecificFailureConnectionReason()
- le_mdc_GetPlatformSpecificDisconnectionCode()

Fix:
Prevent a call to a NULL pointer inside these APIs.
Note that these APIs are deprecated and replaced by new ones.
They won't be maintained in the future and should not be used.

Resolves: LE-11026
Change-Id: I5dba508cd69246ad38184c101a55b1cb3f34281c