Fix gdb permission issues when trying to debug a sandboxed application
The gdbCfg tool does not manage cases when the device reboots or legato restarts.
It assumes that the environment is configured correctly if a certain file exist.
However, when a reboot/restart occurs, the file can exist but the mount point
and smack rule need to be set up again.

Resolves: LE-11274
Change-Id: I1588159d21e4bb072a23cbd681388d6d8a566abb