[DCS] Prevent reporting an outdated connection event
Issue:
Sometimes, Firmware reports two cellular notifications: a connected state
and a disconnected state to the QMINewSessionStateHandler() successively.
The second notification deletes the wdsClient and cleans up.
In the meantime, the first notification is being reported to Legato higher
layer. When it reaches the APIs, the state is already false and we are
unable to get the name of the interface.

Fix:
If unable to gather the interface name, due to a disconnected state,
do not report the connection event to Legato apps since it's outdated.

Resolves: LE-11280
Change-Id: Icd7410ebbd4febb7fe85cf78fd712d2214132c67