Fix for SMACK and IMA enabled systems
The recent support for SMACK onlycap that changed supervisor to run as
admin has caused a regression in a SMACK and IMA enabled system. Applications
are unable to start because they can't access the keys imported by the supervisor
as admin. The fix is to import the key with the '_' label like before so that
everyone has read access to them.

Resolves: LE-11352
Change-Id: Ifb5e6bdc04b05cb4e470b747b1d59cd2ff5787b3