Fix build error for bundled files override case.
If adef and cdef bundle files with same names and same target
location, the build fails.
According to documentation, in case of such conflict,
adef-bundled file must take precedence over cdef-bundled file,
and the build should succeed.
The fix is to revert the check for the name conflict that was
introduced in LE-7477

Resolves: LE-8216
Change-Id: I57ecacae6d1457242cf1073f43065701d282d949