[Docs] Add 18.10 Release Notes
The format for the Release Info page is
changing for the 18.10 release and the release
information table is being moved to legato.io/releases.html.
the aboutReleaseInfo page will now only contain the release
notes related to the current release with the releases.html
page containing the master list of all previous releases.

This change will be backported to all previous releases at
a future date.

Resolves: LE-11491
Change-Id: I8983a9343c963b6c43727ff7939a8fa8ed3fb589