Provide --host configure option for kconfig-frontends
With Yocto toolchain, --host option is required for configure script
otherwise it cannot find libtool.

This is a change required with the version of autotools provided by
Yocto.

Also in the case where /opt/swi/<version>/environment-* is sourced,
few variables are set to target tools (CC, LD, OBJDUMP, ...).
We need to make sure that these tools are not used when building
the native package kconfig-frontend.

Resolves: LE-11563
Change-Id: Ia8fa714f88a1d6726de51c22f734d8b8ad8f7fea