Fix bundling of directories into the same destination path
When bundling multiple directories into the same location, the mktool would
only insert the content of one of the directories. The FileSystemObject_t set
that manages each directory bundled into the app currently sees two object that
have the same destination path as the same object. Two objects are only the same
when both source and destination path are the same.

Resolves: LE-8217
Change-Id: I736425ba82a1de5786603b7c0aeb214ae430444d