Provide sdef 'commands' wrapper
Executing app tools exposed through the sdef 'commands:' directly
results in them inheriting the label of the process that executes
them.

We could eventually use SMACK64EXEC to set the label that they should
be executed at, but in that case we would also need to have proper rules
between stdout/stderr and the label that the process is running at.
Given that the console is typically 'admin' and the app is typically
'app.tools' we would need to let 'app.tools' write to 'admin', which is
not so great.

By wrapping it with 'app runProc', we are executing the tool within the
context of the app.
When it starts, 'app' has label 'admin' so it has the rights to write to
the console. Then the process changes its own label to the 'app.tools'
label, but since the pipe to stdout/stderr is already established, the
tool ends up being able to print properly.

The drawback of this solution is that execution of tools is slower.

Resolves: LE-7010
Change-Id: I929e4c916f2f72cd6e7c034c4657fe074ef895b0