The module cannot go to sleep due to unprocessed event.
The problem is that start.c is loading the Modem PA and not
processing all the event generated by the modem. As a result,
"eventfd" is permanently present in the wakeup sources, and
this prevents module from suspending.
The fix is to spawn a dedicated thread that performs loading
the modem PA, and processes events indefinitely.

Resolves: LE-11212
Change-Id: Ibcfc29f01b4911b069a431e6bd8e56b1fc289d2a