Fix update of an IMA and smack onlycap system
Performing an update of an IMA and smack only system will fail. It is failing
for two different reasons:
1) Apps cannot communicate with the service directory
2) Apps cannot read ld.so.cache

Objects in tmpfs need to be relabeled to address this failure. It was working
on some platforms because there was a SMACK issue where objects created in tmpfs
will always have the '*' label.

Resolves: LE-11500
Related: LXSWIREF-1170
Change-Id: Ibbed345d1fdcd09b52b33df24dcac2cc5eeea115