[virt] Make tools copy optional
During generation of the Docker image, it fails if it can't copy the tools.
Therefore, making this copy optional.

In case of a Yocto build, the Legato build directory is not necessarily
located in $LEGATO_ROOT.

Also providing an optional LEGATO_BUILD_DIR env variable to provide an
eventual build directory to the script.

Also removing the dependency on jq by using docker inspect --format

Resolves: LE-11741
Change-Id: Ia635fde5604174616f6350d6e6b777f581000aa5