Prevent supervisor from utilizing most of the CPU
In the case where an application is in a 'D' process state (uninterruptible),
supervisor will continually try to kill the process until most of the CPU
is used up. This fix is prevent supervisor from entering this state by checking
the processes state before attempting to kill it.

Resolves: LE-11842
Related: LE-11760
Change-Id: I958e2286e5458c8483064784bfa50e589620e9ad