Legato does not start after a rollback of an application update
Address issue where a legato system is unable to start if a rollback is
triggered after an application update. The smack label of the snapshotted
is set incorrectly. Therefore, when a rollback occurs, the system is unable
to access the content of the previous system and start properly.

Resolves: LE-11904
Related: LE-11885
Change-Id: I458652310dd134bb9540a3e808b8bdd478013dc1