Fix host tool 'fwupdate download' exiting with 1 on target reboot
Adding logs to the fwupdate target tool and fwupdate host tool to better
understand the part of the process that they are in, and once we reached
the last part before the reboot, let the host tool end the command with
exit 0 even though the SSH connection has been interrupted.

Also adding some debounce mechanism to avoid calling le_fwupdate_ConnectService
twice.

Resolves: LE-11944
Change-Id: I7a4137f211a23f568cd0782f3323502258ab1a8a