Config Tree optimizations
1) When inserting new item, there is a check whether the node
with such name exists (and this check is expensive, as it iterates
through the linked list). The problem is, this check is performed
twice for each single item; this is optimized by removing redundant
iterations.
2) When checking whether the child node with a given name exists,
the string copying and comparison is performed. This is optimized
by comparing the hashes of the node names, and comparing the
strings only if the hash matches (to cover the possibility of
collision).

Resolves: LE-11925
Change-Id: I688a0b3349684eaa113bf048769fe5f56c2bd4f6