Update SMACK rules
Updating some of the SMACK rules that enabled '_' processes full access
to 'admin' and 'framework'. This requires relabelling some of the shared
resources between legato and the system.

In addition, only allow read access from framework to admin. This requires
reworking the 'sdir list' to write the data to a different fd (stdout).

Resolves: LE-12098
Change-Id: I4f94b5564d29e42903b95fb63518cffa6f5dd64c