Determine dynamically if getrandom is available on build-time glibc >= 2.25
With Yocto 2.5, we are starting to use glibc 2.25 which provides the
getrandom function. However this adds a dependency to that version of
glibc, which makes liblegato.so not run-time compatible with older
versions.

So basically this prevents a liblegato.so built with Yocto 2.5 to run
on Yocto 2.2.

Since we would prefer to use getrandom, this tries to load the symbol
dynamically from the libc at runtime.

Resolves: LE-12152
Change-Id: I5b2de5cc172a8ce176f62cbf6f9e81bccf2e0f0a
Signed-off-by: Bertrand Roussel <broussel@sierrawireless.com>
(cherry picked from commit 9b5781ca3047ad843e5c01fe803078ab4e089789)