Fix fork without exec in network time handling
The network time handling code was calling fork() followed by system(),
which is not safe.  In addition, by not calling an exec variant all
file descriptors were inherited by the child process.  Calling execl()
causes any files opened with O_CLOEXEC (such as the QMI wakeup pipes
which were causing this bug) to close, preventing the terminating child
from potentially pulling the rug out from under the parent.

Originally, this code was paired with a QMI change to set the O_CLOEXEC
flag, but due to the difficulty in getting that integrated this was
changed to use a new le_proc_Execute() function which can handle
closing all of the open file handles when the child is forked.

This also fixes decoding of library paths in dump2sym.

Resolves: LE-12243
Change-Id: If1b0b03579c14fcc53ee499475023ac28669650f