[sbhelper] Update to openat hook and ignore directory access
Issue Description:
Sbhelper fails to import directory '/data/le_fs/' and ignore
to import file which is used by system call open().

Root Cause:
If the app attempts to access the resource which is a directory,
SBhelper will fail.
Sbhelper can't hook system call open() because Glibc version 2.26
switched to implementing open with openat. It uses __NR_openat as
default syscall for open{64}.

Solution:
The fix is to check whether the resource is a directory, and if
it is, silently ignore it and proceed without aborting the sbhelper.
Add syscall __NR_openat in sbtrace which is used by sbhelper.

Resolves: LE-13669 LE-11395
Change-Id: Id8e71ab9383789c82540feb35c6d051994f0b600
Signed-off-by: Rinne Zhang <RiZhang@sierrawireless.com>
(cherry picked from commit cf1aa016496776d68e089a59f75c373b842bb533)