Fix string handling warnings from GCC 8.2.
* components/atServices/atClient/le_atClient.c
(le_atClient_SetIntermediateResponse, le_atClient_SetFinalResponse):
Here, we can replace a strncpy by a memcpy. The reason is that the
destination array is a C99 VLA that is has been allocated to match the
size of the source operand. This is exactly like a strdup, except into
the stack rather than heap.

* components/atServices/atServer/bridge.c (AtCmdHandler):
Initialization of local array to zero bytes done with initializer
instad of memset. We shrink the size of cmdName by a few bytes,
and use sizeof cmdName to refer to its size, instead of proliferating
a preprocessor constant. The reason we shrink cmdName is that it's
used in some snprintf formatting calls which go into a destination
buffer that is only as large as cmdName, so that the operations may be
truncated.

* components/cellNetService/cellNetServiceServer.c
(le_cellnet_GetSimPinCode): We replace the incorrect strncpy(x, y,
strlen(y)) pattern which fails to null terminate with le_utf8_copy.
This returns Legato result which we can log and pass on.

* components/dataConnectionService/dcs/dcs_db.c
(le_dcs_CreateChannelDb): strncpy with length obtained from source
operand is replaced with le_utf8_Copy with size obtained from
destination.

* components/dataConnectionService/dcsDaemon/dcsServer.c
(TryStartTechSession): More instances of strncpy to le_utf8_Copy.

* components/dataConnectionService/dcsWifi/dcsWifi.c
(DcsWifiCreateConnDb): Likewise.

* components/modemServices/modemDaemon/le_sms.c (le_sms_SetText):
Likewise.

* framework/daemons/linux/start/start.c (LoadModemPA): Simplified
the calculation of the modem path into a single concise snprintf call,
whose return value we use as the length check.  Since snprintf doesn't
allow the destination buffer to overlap with any of the source strings
(undefined behavior), we use a separate buffer for obtaining the
symlink. Adding logs for all the bail conditions: cannot read symlink,
symlink doesn't contain dot, total path is too long.

* framework/daemons/linux/supervisor/kernelModules.c (ExecuteCommand):
Here, the argv[] to be executed is being logged via LE_INFO message,
and for that reason, the function joins the arguments into a single
string with spaces. If there are too many arguments, then "..." is
added. I'm rewriting the loop to avoid strcat. Everything is done
with snprintf, and made a bit more concise also.

Resolves: LE-12273
Change-Id: I616a43fb5654e1ab997999ea99c944e1fd8ecade