Add support to generate multiple ko files from a single MDEF
This patch provides capability to generate multiple kernel object
files (.ko) from a single module definition file (MDEF).

Introduce section 'kernelModule:' in MDEF, which supports sub-sections:

name: Name of the kernel module.
sources: The list of source files for the module.
requires: kernelModules: Dependency for the module.

Update the parser and modeller to support the new sections.
Update the config generator code to print the sub modules to
the config tree.

Resolves: LE-11968

Change-Id: I2324611079a3b2cc5eaee812ae9366a013975cd9