Implement Legato RPC feature (Drops # 1, 2 and 3)
This commit adds infrastructure for supporting the RPC Proxy,
an application (or library) that is central for hosting multi-
host RPC Bindings.

- Adds Tag-ID support to the serialized pack data.
- Adds KConfig RPC feature flags.
- Splits the RPC Proxy meta-data into separate Static and Dynamic
configuration definitions.

- Adds support for handling the split in RPC Proxy meta-data,
including new code for retrieving configuration fom the ConfigTree.

- Extends the RPC Proxy infrastructure to support Session and Service
creation on a LINUX Legato system, that rely on the presence of the
Service-Directory to bind Client and Server.

- Updates Legato documentation to include RPC as an EXPERIMENTAL
feature.

Resolves: LE-10096 LE-11512 LE-11507 LE-11509 LE-11506 LE-11510 LE-11513 LE-11532 LE-11524 LE-11525 LE-12202
Change-Id: I46d8594b6b203fdd06021c3b33539d0511e9b51c