Find more recent versions of Yocto toolchains
The path was hardcoded to /opt/swi/y22 and /opt/swi/y22-ext, even for
more recent versions of Yocto such as 2.5
Since we are now generated the path dynamically on the Yocto side,
we want to make the Legato side more flexible.

Updating the findtoolchain script to handle more recent versions
dynamically.

This only has effect on dev versions of the Yocto toolchains, since
versionned ones are using /opt/swi/<tag>, so
/opt/swi/LXSWI2.5-5.0+wp76xx
OR
/opt/swi/SWI9X07Y_02.14.04.00

Resolves: LE-12586
Change-Id: Ie0e9421e5a7606d4a4a58b31b78b4b62571af9e8