Add support for externalBuild section in MDEF
Allow support for external build process in module definition file.
A common use case is to specify a Makefile and run make to build a module.
The Makefile must include everything required to build a kernel module.

If externalBuild section is specified, source code file paths is not
allowed in sources section.

externalBuild:
{
    "make -f $CURDIR/Makefile"
}

The generated module object is then listed in the preBuilt section.

Resolves: LE-11984
Change-Id: I8ef8f1c6559ba56c93c5a06bca6be0bb319feb03