Detect if __attribute__((error)) is supported.
Some versions of clang do not support __attribute__((error)).  If
that's the case, use an alternate implementation of LE_FULL_API, or
disable it completely if no suitable implementation is possible.
This is a convenience feature only -- in worst case the user will
get a link error instead of a friendlier compile error.

Resolves: LE-12783
Change-Id: Iada90d46852f3c42a3d5567340444133829f6f81