Remove private liblegato functions from generated code
Generated code should not rely on any internal headers or functions of
liblegato.

To avoid passing in the entire function prototype in a -D option at the
command line, use macros to build up the prototype and similar required
values from the component name.

The prototype of the init functions was changed to avoid having to
include a private liblegato header in the generated code.

Resolves: LE-12588
Change-Id: Ib85ba6536ca0a1d6328135ad5210a7a8f9d96cfe