Update MDC's local copy of modemData before setting a new parameter in it back into modem
Contents of modemData in the modem might change without updating the same in MDC in Legato.
Then, when a modemData change is set via MDC, it writes back into the modem what it knows
which might be outdated. The fix is to update the modemData in MDC before writing any
new parameter change back in.

Resolves: LE-11783
Related: LE-11307 LEAR7M-429
Change-Id: I3fb9ec7bbc7f837b45a9c2c6ed0af71809c8df96