Fix RPC RTOS Support
Recent changes to enable RPC on LINUX in the
19.04 Release have resulted in some
incompatibilities with supporting RPC on
RTOS (HL78).

These changes address the following issues:

- Harmonizing RPC configuration configTree
support between the Linux and HL78
platforms.

- Update the RPC Configuration tool to
accommodate the harmonization changes being
made to the RPC configTree support.
As a part of this, a new RPC Configuration
'shim' tool has been added to the HL78 targetTools.

- Extended the RPC Configuration tool to
display status information for bindings
and links.

- Initialize le_comm networkSocket component
in the le_comm_Create() function, if not
yet initialized.

- Fix issue in the RPC Proxy Repack functions
used on the HL78 platform.

- Initialize the RPC Service memory pools to
include the LE_MSG_LOCAL_HEADER_SIZE.

- Fix issue with the handling of the RPC Proxy DeleteService(),
affecting the RTOS (HL78).

Resolves: LE-12909
Change-Id: Id9d2bb1d60e91e6e999f09e73b515e75e7c07500