Fix Bug for FTP When Downloading Files Larger Than 10K
Make SinkData not stuck in spin if new event cannot be allocated and return 0 instead.
Also change the EventDestructor so the receive resume function is called only if
the event being destroyed has a buffer.

Adding checks after an attempt to create new events.
Fixing header of the NewEvent function.

Resolves: LE-12660
Change-Id: I76d3b1cbd9aecd153bfec859dce45b2db9fc91b7