Prototype of MCU <-> MAP communication
These changes:

- Extend the Ninja build file dependency statements
for the RTOS file rpcService.c to ensure the generated interface headers
have been created when it is being built.

- Fix an issue affecting the argument handling in the StartProc() function
in the RTOS microSupervisor.

- Refactor the RPC Configuration 'shim' tool to directly pass the command-line
arguments to the microSupervisor.

- Marks unused parameters, using LE_UNUSED.

- Adds implementation of le_msg_RequestResponse in the RTOS messaging.c file.

- Splits the RPC Proxy's handling of the Dynamic Run-time
configuration into two distinct cases: (1) ConfigTree, and (2) Statically-defined.
For the statically-defined case, the function rpcProxyConfig_LoadBindings()
must now be provided by a component other than the RPC Proxy.  For the MCU build,
which requires statically-defined run-time configuration, this is provided
by the SWI component, rpcStaticConfig.

- Defined two new KConfig integer flags for RPC performance tweaking,
affecting the number of RPC services and the number of RPC "out" parameters.

- Introduces the RPC Message Re-assembly State-Machine logic.   This is needed in
order to deal with use-cases in which the underlying communication does not
guarantee receiving complete message packets.

-- Fix an issue affecting the Server-Response memory buffer handling.

Resolves: LE-12984
Change-Id: Idfebd6dd95fab0a5dc29575da51b727fc7dcaaf1